Create sqlite-to-mariadb-alt.pl
This file is for when your DB is for TS3 and you have backslashes in the channel paths which causes an error when you try to import the DB after running the mariadb-ts3-fixes.sh file. You will need to re-export the DB, then run this and then finally the TS3 fixes.